submit java-oop-ru/interfaces

+++ java-oop-ru/interfaces/src/main/java/exercise/App.java
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 import java.util.ArrayList;
 public class App {
     public static List<String> buildApartmentsList(List<Home> apartments, int number) {
-        apartments.sort((ap1, ap2) -> Integer.compare(ap1.getArea(), ap2.getArea()));
+        apartments.sort((ap1, ap2) -> Double.compare(ap1.getArea(), ap2.getArea()));
         List<String> result = new ArrayList<String>();
 
         for (int i = 0; i <= number; i++) {
